more test and fixes

+++ src/main/kotlin/com/funcorp/bandit/algorithm/ICalculateScoreStrategy.kt
@@ -1,5 +1,5 @@
 package com.funcorp.bandit.algorithm
 
 interface ICalculateScoreStrategy {
-    fun calculateScore(attempts: Int, score: Double, reward: Double): Double
+    fun calculateScore(attempts: Long, score: Double, reward: Double): Double
 }
